Move onebp_variants function to utils
