+++ thapbi_pict/utils.py
@@ -10,6 +10,54 @@ from Bio.Data.IUPACData import ambiguous_dna_values
 from Bio.SeqIO.FastaIO import SimpleFastaParser
 
 
+def onebp_variants(seq):
+    """Generate all 1bp variants of the sequence (substitution, deletion or insertion).
+
+    Assumes unambiguous IUPAC codes A, C, G, T only.
+    """
+    seq = seq.upper()
+    variants = set()
+    for i in range(len(seq)):
+        # One base deletion
+        variants.add(seq[:i] + seq[i + 1 :])
+        for s in "ACGT":
+            # One base substitions
+            variants.add(seq[:i] + s + seq[i + 1 :])
+            # One base insertions
+            variants.add(seq[:i] + s + seq[i:])
+    for s in "ACGT":
+        # One base "insertion" at the end
+        variants.add(seq + s)
+    variants.remove(seq)
+    return variants
+
+
+assert set(onebp_variants("A")) == set(
+    ["", "C", "G", "T", "AA", "CA", "GA", "TA", "AC", "AG", "AT"]
+)
+assert set(onebp_variants("AA")) == set(
+    [
+        "A",
+        "CA",
+        "GA",
+        "TA",
+        "AC",
+        "AG",
+        "AT",
+        "AAA",
+        "CAA",
+        "GAA",
+        "TAA",
+        "ACA",
+        "AGA",
+        "ATA",
+        "AAC",
+        "AAG",
+        "AAT",
+    ]
+)
+
+
 def expand_IUPAC_ambiguity_codes(seq):
     """Convert to upper case and iterate over possible unabmigous interpretations.
